[editor] websites view for editors

+++ app/controllers/websites.js
@@ -9,18 +9,26 @@ angular.module('MyApp')
         }
 
       $scope.newWebsite = function () {
+          if(!$rootScope.isAdmin())
+              return;
           $scope.currentWebsite = {id:0, name:'', url: '', git_url:'', editors: []};
       }
 
       $scope.updateWebsite = function (index) {
+          if(!$rootScope.isAdmin())
+              return;
           $scope.currentWebsite = $scope.websites[index];
       }
 
       $scope.deleteWebsite = function (index) {
+          if(!$rootScope.isAdmin())
+              return;
           $scope.currentWebsite = $scope.websites[index];
       }
 
       $scope.deleteCurrentWebsite = function () {
+          if(!$rootScope.isAdmin())
+              return;
           Websites.deleteWebsite($scope.currentWebsite.id).then(function (response) {
               $scope.messages = {
                   success: [{msg:'Website deleted'}]
@@ -34,6 +42,8 @@ angular.module('MyApp')
       }
 
       $scope.saveCurrentWebsite = function () {
+          if(!$rootScope.isAdmin())
+              return;
           var func = null;
           if($scope.currentWebsite.id == 0)
               func = Websites.postWebsite($scope.currentWebsite);
@@ -55,8 +65,10 @@ angular.module('MyApp')
       function getWebsites(){
           Websites.getWebsites().then(function (response) {
               var i;
-              for(i = 0; i<response.data.websites.length; i++)
-                  response.data.websites[i].editors = $rootScope.libs.utils.pluck(response.data.websites[i].editors, 'id');
+              // we have editors only if we are admin
+              if($rootScope.isAdmin())
+                  for(i = 0; i<response.data.websites.length; i++)
+                      response.data.websites[i].editors = $rootScope.libs.utils.pluck(response.data.websites[i].editors, 'id');
               $scope.websites = response.data.websites;
           }).catch(function (response) {
               $scope.messages = {
@@ -66,6 +78,8 @@ angular.module('MyApp')
       }
 
       function getEditors(){
+          if(!$rootScope.isAdmin())
+              return;
           Editors.getEditors().then(function (response) {
               $scope.editors = response.data.editors;
           }).catch(function (response) {
